requirements.txt
Django==1.8.3
django-jsignature==0.8
django-signature==0.3.1
django-widget-tweaks==1.4.1
mysqlclient==1.3.6
Pillow==2.9.0
python-signpad2image==0.0.2
six==1.9.0
wheel==0.24.0
